services/tg_bot/requirements.txt
aerich
aiogram
fastapi
tortoise-orm
tortoise-orm[asyncpg]
PyJWT
redis
pytest
python-dotenv
passlib
python-multipart
email-validator
bcrypt
cryptography
uvicorn
Tortoise